requirements.txt
fastapi~=0.115.12
uvicorn[standard]~=0.34.2
langgraph~=0.4.3
langchain
langchain-openai~=0.3.16
pydantic[dotenv]~=2.11.4
sqlalchemy[asyncio]~=2.0.41
asyncpg
alembic~=1.15.2
python-dotenv~=1.1.0
psycopg2-binary # Often needed by Alembic even with asyncpg
mcp # MCP Core SDK
langchain-mcp-adapters # Adapters for Langchain/LangGraph 
langchain-core~=0.3.59
pydantic-settings~=2.9.1
langchain-google-genai~=2.1.4